Add keyword based resolder functionality

+++ unit.go
@@ -27,6 +27,13 @@ func (u *Unit) Start(ctxt *ishell.Context) error {
 	return slot.Start(u, ctxt)
 }
 
+func (u *Unit) Restart(ctxt *ishell.Context) error {
+	if err := u.Stop(ctxt); err != nil {
+		return err
+	}
+	return u.Start(ctxt)
+}
+
 func (u *Unit) OutputFile() (*os.File, error) {
 	home := os.Getenv("HOME")
 	if len(home) == 0 {
@@ -64,8 +71,15 @@ func (u *Unit) Stop(ctxt *ishell.Context) error {
 
 	u.Status.shutdownRequested.Set()
 
+	u.Status.Lock()
+	defer u.Status.Unlock()
 	// TODO(ttacon): move to refactored function
 	if u.Status.CurrentSlot.Provider.Type == "bash/local" {
+		// It's possible to be beaten here by the goroutine that is
+		// waiting on the process to exit, so safety belts!
+		if u.Status.Cmd == nil {
+			return nil
+		}
 
 		if err := u.Status.Cmd.Process.Kill(); err != nil {
 			return err
